Reorganize Structure of project to be more scalable
